Moved Builder source code to a new folder.
