+++ Source/tools/builder/code/analyzer.cpp
