+++ Source/tools/builder/code/analyzer.h
@@ -2,7 +2,7 @@
 #ifndef DSR_BUILDER_ANALYZER_MODULE
 #define DSR_BUILDER_ANALYZER_MODULE
 
-#include "../../DFPSR/api/fileAPI.h"
+#include "../../../DFPSR/api/fileAPI.h"
 #include "builderTypes.h"
 
 using namespace dsr;
