+++ Source/tools/builder/code/builderTypes.h
@@ -2,7 +2,7 @@
 #ifndef DSR_BUILDER_TYPES
 #define DSR_BUILDER_TYPES
 
-#include "../../DFPSR/api/stringAPI.h"
+#include "../../../DFPSR/api/stringAPI.h"
 
 using namespace dsr;
