+++ Source/tools/builder/code/generator.h
@@ -2,7 +2,7 @@
 #ifndef DSR_BUILDER_GENERATOR_MODULE
 #define DSR_BUILDER_GENERATOR_MODULE
 
-#include "../../DFPSR/api/fileAPI.h"
+#include "../../../DFPSR/api/fileAPI.h"
 #include "builderTypes.h"
 
 using namespace dsr;
