+++ Source/tools/builder/code/main.cpp
@@ -65,7 +65,7 @@ Project files:
 		* Optimization, a natural integer specifying the amount of optimization to apply.
 */
 
-#include "../../DFPSR/api/fileAPI.h"
+#include "../../../DFPSR/api/fileAPI.h"
 #include "Machine.h"
 #include "expression.h"
 #include "analyzer.h"
